% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{asymp_cov}
\alias{asymp_cov}
\title{Asymptotic standard errors for correlation matrices.}
\usage{
asymp_cov(R, X = NULL, eta = 1, type = "normal")
}
\arguments{
\item{R}{Correlation matrix.}

\item{X}{Optional raw data matrix.}

\item{eta}{Skewness parameter for elliptical data distributions.}

\item{type}{Type of random deviates: "normal", "elliptical" or "general".}
}
\value{
The asymptotic covariance matrix of \code{R}.
}
\description{
Get the asymptotic standard errors of correlation matrices of normal or arbitrary random deviates.
}
\details{
If \code{type = "normal"}, the calculation assumes that the raw data follows a multivariate normal distribution.
If \code{type = "elliptical"}, the calculation assumes that the raw data follows an elliptical distribution with skewness parameter \code{eta}.
If \code{type = "general"}, no assumption is made but need to provide the raw data via the \code{X} argument.
}
\references{
M.W. Browne and A. Shapiro (1986). The asymptotic covariance matrix of sample correlation coefficients under general conditions. Linear Algebra and its Applications, 82, 169-176. https://doi.org/10.1016/0024-3795(86)90150-3
}
