% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{bifactor}
\alias{bifactor}
\title{Fit an exploratory bifactor model}
\usage{
bifactor(
  R,
  n_generals,
  n_specifics,
  method,
  rotation,
  init = NULL,
  normalize = FALSE,
  gamma = 0,
  epsilon = 0.01,
  k = 0,
  w = 1,
  bifactor_method = "SL",
  SLiD_max_iter = 10L,
  cutoff = 0.2,
  PhiTarget = NULL,
  PhiWeight = NULL,
  random_starts = 1L,
  cores = 1L,
  efa_max_iter = 10000L,
  efa_factr = 1e+07,
  m = 5L,
  rot_max_iter = 10000L,
  rot_eps = 1e-05,
  verbose = TRUE
)
}
\description{
Fit an exploratory bifactor model
}
