% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{check_deriv}
\alias{check_deriv}
\title{Check the derivatives and differentials of rotation criteria.}
\usage{
check_deriv(L, Phi, dL, dP, rotation = "oblimin", projection = "oblq",
Target = NULL, Weight = NULL, PhiTarget = NULL, PhiWeight = NULL,
blocks = NULL, blocks_list = NULL, block_weights = NULL,
oblq_blocks = NULL, between_blocks = "none", gamma = 0,
epsilon = 0.01, k = 0L, w = 1, alpha = 1, a = 30, b = 0.36)
}
\arguments{
\item{L}{Loading matrix.}

\item{Phi}{Factor correlation matrix.}

\item{dL}{Perturbation for L.}

\item{dP}{Perturbation for Phi.}

\item{rotation}{Rotation criterion. Available rotations: "varimax", "cf" (Crawford-Ferguson), "oblimin", "geomin", "target", "xtarget" (extended target) and "none". Defaults to "oblimin".}

\item{projection}{Projection method. Available projections: "orth" (orthogonal), "oblq" (oblique), "poblq" (partially oblique). Defaults to "oblq".}

\item{Target}{Target matrix for the loadings. Defaults to NULL.}

\item{Weight}{Weight matrix for the loadings. Defaults to NULL.}

\item{PhiTarget}{Target matrix for the factor correlations. Defaults to NULL.}

\item{PhiWeight}{Weight matrix for the factor correlations. Defaults to NULL.}

\item{blocks}{Vector with the number of factors for which separately applying the rotation criterion. Defaults to NULL.}

\item{blocks_list}{List containing the columns to which applying the rotation criterion.}

\item{block_weights}{Vector of weights for each block of factors.}

\item{oblq_blocks}{Vector with the number of factors for each oblique block. E.g.: c(2, 4) means that there are two blocks of oblique factors: one block with 2 factors and another block with 4 factors. Everything else is orthogonal. Defaults to NULL.}

\item{between_blocks}{Available between_blocks: "TL" and "TLM". Defaults to none.}

\item{gamma}{\eqn{\gamma} parameter for the oblimin criterion. Defaults to 0 (quartimin).}

\item{epsilon}{\eqn{\epsilon} parameter for the geomin criterion. Defaults to 0.01.}

\item{k}{\eqn{k} parameter for the Crawford-Ferguson family of rotation criteria. Defaults to 0.}

\item{w}{\eqn{w} parameter for the extended target criterion ("xtarget"). Defaults to 1L.}

\item{alpha}{\eqn{\alpha} parameter for the Tian and Liu between_blocks. Defaults to 1.}

\item{a}{Discrimination parameter for simplix rotation. Defaults to 30.}

\item{b}{Difficulty parameter for simplix rotation. Defaults to 0.36.}
}
\value{
A list with the objective value and the gradients and differentials of L and Phi.
}
\description{
Check the derivatives and differentials of rotation criteria.
}
\details{
None yet.
}
