% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{cv_eigen}
\alias{cv_eigen}
\title{Cross-validated eigenvalues.}
\usage{
cv_eigen(X, N = 100L, hierarchical = FALSE, efa = NULL, cores = 1L)
}
\arguments{
\item{X}{Raw data matrix.}

\item{N}{Number of cross-validated samples.}

\item{hierarchical}{Logical indicating whether a second cross-validated eigenvalues estimation should be performed from the factor scores obtained after a first factor analysis analysis.}

\item{efa}{A list of arguments to pass to \code{efast} when \code{hierarchical = TRUE}.}

\item{cores}{Number of cores to perform parallel computations.}
}
\value{
A list with the cross-validated eigenvalues and the estimated dimensionality.
}
\description{
Estimate cross-validated eigenvalues and the dimensionality using the Kaiser's rule.
}
\details{
None yet.
}
\references{
Chen F., Roch S., Rohe K., Yu S (2021). Estimating Graph Dimension with Cross-validated Eigenvalues, arXiv. https://arxiv.org/abs/2108.03336
}
