% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{efast}
\alias{efast}
\title{Fast exploratory factor analysis}
\usage{
efast(
  R,
  n_factors,
  method = "minres",
  rotation = "oblimin",
  init = NULL,
  Target = NULL,
  Weight = NULL,
  PhiTarget = NULL,
  PhiWeight = NULL,
  oblique_indexes = NULL,
  normalize = FALSE,
  gamma = 0,
  epsilon = 0.01,
  k = 0,
  w = 1,
  random_starts = 1L,
  cores = 1L,
  efa_max_iter = 10000L,
  efa_factr = 1e+07,
  m = 5L,
  rot_max_iter = 10000L,
  rot_eps = 1e-05
)
}
\description{
Fast exploratory factor analysis
}
