% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Compute fit measures for exploratory factor models.}
\usage{
fit(model, nobs = NULL)
}
\arguments{
\item{nobs}{Sample size. Defaults to NULL.}

\item{model}{Object of class efa or cfa.}
}
\value{
Vector of fit measures.
}
\description{
Compute fit measures for exploratory factor models.
}
\details{
\code{fit}... to be explained
}
\author{
Vithor R. Franco & Marcos Jiménez
}
