% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMeasures.R
\name{fitMeasures}
\alias{fitMeasures}
\title{Compute fit measures for exploratory factor models.}
\usage{
fitMeasures(efa, nobs = NULL)
}
\arguments{
\item{efa}{Object of class efa.}

\item{nobs}{Sample size. Defaults to NULL.}
}
\value{
Vector of fit measures.
}
\description{
Compute fit measures for exploratory factor models.
}
\details{
\code{fitMeasures}... to be explained
}
\author{
Vithor R. Franco & Marcos Jiménez
}
