% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscores.R
\name{fscores}
\alias{fscores}
\title{Compute factor scores}
\usage{
fscores(fit, scores = NULL, method = "regression")
}
\arguments{
\item{fit}{object of class `efa` or `bifactor`.}

\item{scores}{Matrix of raw scores.}

\item{method}{Method to compute the factor scores.}
}
\value{
List...
}
\description{
Compute the factor scores from an exploratory factor model
}
\details{
...
}
\author{
Marcos Jiménez & Vithor R. Franco
}
