% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifactor.R
\name{gen}
\alias{gen}
\title{title: Square a number}
\usage{
gen(N, items, reps)
}
\arguments{
\item{N}{A number of observations per dataset to generate}

\item{items}{A number of items per dataset to generate}

\item{reps}{Number of replications}
}
\value{
Rturns Correlations between thetas, and RMSEA values for EAP scores
}
\description{
Takes in any numeric value and squares it
}

