% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phi_target.R
\name{get_phi_target}
\alias{get_phi_target}
\title{Get a target for the correlation matrix}
\usage{
get_phi_target(n_generals, specifics_per_general, g_rho = TRUE, s_rho = FALSE)
}
\description{
Get a target for the correlation matrix
}
