% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{get_target}
\alias{get_target}
\title{Get a target from a loading matrix.}
\usage{
get_target(loadings, Phi = NULL, cutoff = 0)
}
\arguments{
\item{loadings}{A matrix of loadings.}

\item{Phi}{A correlation matrix among the factors. Defaults to NULL.}

\item{cutoff}{The cut-off used to create the target matrix. Defaults to 0.}
}
\value{
A target matrix.
}
\description{
Get a target for the loading matrix using a custom or empirical cut-off.
}
\details{
If \code{cutoff} is not 0, loadings smaller than such a cut-off are fixed to 0. When \code{cutoff = 0}, an empirical cut-off is used for each column of the loading matrix. They are the mean of the one-lagged differences of the sorted squared normalized loadings. Then, the target is determined by fixing to 0 the squared normalized loadings smaller than such cut-offs.
}
\references{
Garcia-Garzon, E., Abad, F. J., & Garrido, L. E. (2019). Improving bi-factor exploratory modeling: Empirical target rotation based on loading differences. Methodology: European Journal of Research Methods for the Behavioral and Social Sciences, 15(2), 45–55. https://doi.org/10.1027/1614-2241/a000163

Jiménez, M., Abad, F.J., Garcia-Garzon, E., Garrido, L.E. (2021, June 24). Generalized exploratory bi-factor Modeling. Under review. Retrieved from https://osf.io/7aszj/?view_only=8f7bd98025104347a96f60a6736f5a64
}
