% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{get_target_with_cutoff}
\alias{get_target_with_cutoff}
\title{Get a target for loadings by specifying a threshold}
\usage{
get_target_with_cutoff(L, cutoff = 0.2)
}
\description{
Get a target for loadings by specifying a threshold
}
