% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{multiple_rotations}
\alias{multiple_rotations}
\title{Perform multiple rotations to ensure a global minimum}
\usage{
multiple_rotations(
  loadings,
  rotation,
  Target,
  Weight,
  Phi_Target,
  Phi_Weight,
  gamma = 0,
  epsilon = 0.01,
  k = 0,
  w = 1,
  random_starts = 10L,
  cores = 1L,
  eps = 1e-05,
  max_iter = 10000L
)
}
\description{
Perform multiple rotations to ensure a global minimum
}
