% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bifactor.R
\name{print.bifactor}
\alias{print.bifactor}
\title{Print summary information from exploratory factor models.}
\usage{
## S3 method for class 'bifactor'
print(bifactor, nobs=NULL, ...)
print.bifactor(bifactor, nobs=NULL, ...)
}
\arguments{
\item{nobs}{Sample size. Defaults to NULL.}

\item{...}{Arguments to be passed to or from other methods.}

\item{efa}{Object of class bifactor}
}
\value{
Matrix of variance accounted for the factors.
}
\description{
Print summary information from exploratory factor models.
}
\details{
to be explained
}
\author{
Vithor R. Franco & Marcos Jiménez
}
