% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.efa.R
\name{print.efa}
\alias{print.efa}
\title{Print summary information from exploratory factor models.}
\usage{
## S3 method for class 'efa'
print(efa, nobs=NULL, ...)
}
\arguments{
\item{efa}{Object of class efa.}

\item{nobs}{Sample size. Defaults to NULL.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
Matrix of variance accounted for the factors.
}
\description{
Print summary information from exploratory factor models.
}
\details{
\code{print.efa}... to be explained
}
\author{
Vithor R. Franco & Marcos Jiménez
}
