% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{random_oblq}
\alias{random_oblq}
\title{Generate random oblique matrices.}
\usage{
random_oblq(p, q)
}
\arguments{
\item{p}{Number of rows.}

\item{q}{Number of columns. Should not be greater than p.}
}
\value{
An oblique matrix with normally distributed data.
}
\description{
Generate random oblique matrices from a standard normal distribution.
}
\references{
Jiménez, M., Abad, F.J., Garcia-Garzon, E., Garrido, L.E. (2021, June 24). Exploratory Two-tier Modeling. Under review. Retrieved from osf.io/7aszj
}
