% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{retr_oblq}
\alias{retr_oblq}
\title{Retraction of a matrix onto the oblique manifold.}
\usage{
retr_oblq(X)
}
\arguments{
\item{X}{A matrix.}
}
\value{
An oblique matrix.
}
\description{
Transform a matrix into an oblique matrix.
}
\references{
Jiménez, M., Abad, F.J., Garcia-Garzon, E., Garrido, L.E. (2021, June 24). Exploratory Two-tier Modeling. Under review. Retrieved from osf.io/7aszj
}
