% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{retr_poblq}
\alias{retr_poblq}
\title{Retraction of a matrix onto the partially oblique manifold.}
\usage{
retr_poblq(X, oblq_factors)
}
\arguments{
\item{X}{A matrix.}

\item{oblq_factors}{A vector with the number of factors for each oblique block. E.g.: c(2, 4) means that there are two blocks of oblique factors: one with 2 factors and another with 4 factors. Everything else is orthogonal.}
}
\value{
A partially oblique matrix.
}
\description{
Transform a matrix into a partially oblique matrix.
}
\examples{

X <- replicate(8, rnorm(8))
retr_poblq(X, c(2, 3, 3))

}
\references{
Jiménez, M., Abad, F. J., Garcia-Garzon, E., & Garrido, L. E. (2023). Exploratory Bi-factor Analysis with Multiple General Factors. Multivariate behavioral research, 1–18. Advance online publication. https://doi.org/10.1080/00273171.2023.2189571
}
