% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{se}
\alias{se}
\title{Standard errors for rotated factor loadings, factor correlations and uniquenesses.}
\usage{
se(n, fit = NULL, R = NULL, Lambda = NULL, Phi = NULL, X = NULL,
 method = "minres", projection = "oblq", rotation = "oblimin",
 Target = NULL, Weight = NULL, PhiTarget = NULL, PhiWeight = NULL,
 blocks = NULL, blocks_list = NULL, block_weights = NULL, oblq_blocks = NULL,
 gamma = 0, k = 0, epsilon = 0.01, w = 1, alpha = 1,
 normalize = FALSE, penalization = "none", type = "normal", eta = 1)
}
\arguments{
\item{n}{Sample size.}

\item{fit}{Optional \code{efast} model.}

\item{R}{Correlation matrix.}

\item{Lambda}{Estimated factor loadings.}

\item{Phi}{Estimated factor correlations.}

\item{X}{Raw data matrix.}

\item{method}{Method used to estimate the factor model: "minres", "pa", "ml", or "minres".}

\item{projection}{Projection used to rotate the factor loadings: "orth", "oblq" or "poblq".}

\item{rotation}{Rotation criterion. Available rotations: "varimax", "cf" (Crawford-Ferguson), "oblimin", "geomin", "target", "xtarget" (extended target) and "none". Defaults to "oblimin".}

\item{Target}{Target matrix for the loadings. Defaults to NULL.}

\item{Weight}{Weight matrix for the loadings. Defaults to NULL.}

\item{PhiTarget}{Target matrix for the factor correlations. Defaults to NULL.}

\item{PhiWeight}{Weight matrix for the factor correlations. Defaults to NULL.}

\item{blocks}{Vector with the number of factors for which separately applying the rotation criterion. Defaults to NULL.}

\item{blocks_list}{List containing the columns to which applying the rotation criterion.}

\item{block_weights}{Vector of weights for each block of factors.}

\item{oblq_blocks}{Vector with the number of factors for each oblique block. E.g.: c(2, 4) means that there are two blocks of oblique factors: one block with 2 factors and another block with 4 factors. Everything else is orthogonal. Defaults to NULL.}

\item{gamma}{\eqn{\gamma} parameter for the oblimin criterion. Defaults to 0 (quartimin).}

\item{k}{\eqn{k} parameter for the Crawford-Ferguson family of rotation criteria. Defaults to 0.}

\item{epsilon}{\eqn{\epsilon} parameter for the geomin criterion. Defaults to 0.01.}

\item{w}{\eqn{w} parameter for the extended target criterion ("xtarget"). Defaults to 1L.}

\item{alpha}{\eqn{\alpha} parameter for the Tian and Liu penalization. Defaults to 1.}

\item{normalize}{Kaiser normalization. Defaults to FALSE.}

\item{penalization}{Available penalizations: "TL" and "TLM". Defaults to none.}

\item{type}{Type of random deviates: "normal", "elliptical" or "general".}

\item{eta}{Skewness parameter for elliptical data distributions.}
}
\value{
A list with the standard errors of the rotated factor loadings, factor correlations and uniquenesses.
}
\description{
Compute the sandwich standard errors of factor loadings, factor correlations and uniquenesses.
}
\details{
Currently, only available for \code{method = minres}.
}
\references{
Zhang G, Preacher KJ, Hattori M, Jiang G, Trichtinger LA (2019). A sandwich standard error estimator for exploratory factor analysis with nonnormal data and imperfect models. Applied Psychological Measurement, 43, 360–373. https://doi.org/10.1177/0146621618798669
}
