% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bifactor.R
\name{sim_bifactor}
\alias{sim_bifactor}
\title{Simulate a bifactor structure}
\usage{
sim_bifactor(
  n_generals,
  specifics_per_general,
  items_per_specific,
  loadings_g = "medium",
  loadings_s = "medium",
  crossloadings = 0,
  pure = FALSE,
  generals_rho = 0,
  specifics_rho = 0
)
}
\description{
Simulate a bifactor structure
}
