% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{sl}
\alias{sl}
\title{Schmid-Leiman Transformation.}
\usage{
sl(X, n_generals, n_groups, cor = "pearson", nobs = NULL, first_efa = NULL,
second_efa = NULL, cores = 1L)
}
\arguments{
\item{X}{Raw data matrix or correlation matrix.}

\item{n_generals}{Number of general factors.}

\item{n_groups}{Number of group factors.}

\item{cor}{Correlation method. Available correlations: c("pearson", "poly"). Defaults to "pearson".}

\item{nobs}{Sample size. Defaults to NULL.}

\item{first_efa}{Arguments to pass to \code{efast} in the first-order factor extraction. See \code{efast} for the default arguments.}

\item{second_efa}{Arguments to pass to \code{efast} in the second-order factor extraction. See \code{efast} for the default arguments.}

\item{cores}{Number of cores for the polychorics estimation.}
}
\value{
\item{loadings}{Loading matrix of the Schmid-Leiman solution.}
\item{first_order_solution}{Object of class \code{efast} with the first-order solution.}
\item{second_order_solution}{Object of class \code{efast} with the second-order solution.}
\item{uniquenesses}{Vector of uniquenesses.}
\item{Rhat}{Correlation matrix predicted by the (hierarchical) model.}
}
\description{
Schmid-Leiman transformation into a bi-factor pattern with one or multiple general factors.
}
\details{
First, a hierarchical factor model is fitted using a second-order factor analysis on the factor correlation obtained from a first-order factor analysis. Then, the item loadings on the general factors are assumed to be the direct effects of the general factors according to such hierarchical model.
On the other hand, the item loadings on the group factors become the originally first-order loadings post-multiplied by the diagonal matrix containing the root of the item uniquenesses.

Obviously, the first-order factor analysis should be oblique to perform a second exploratory factor analysis.

If the second-order solution does not use an orthogonal projection, then the correlation matrix among the general factors for the Schmid-Leiman solution is simply that obtained from such second-order solution.
}
\examples{

\dontrun{
# Simulate data:
sim <- sim_factor(n_generals = 2, groups_per_general = 3, items_per_group = 5)
lambda <- sim$lambda
Target <- ifelse(lambda > 0, 1, 0)

# Target rotation for the first-order efa and oblimin for the second-order efa:
first <- list(rotation = "target", projection = "oblq", Target = Target[, -c(1:2)])
second <- list(rotation = "oblimin", projection = "oblq", gamma = 0)

SL <- <- sl(sim$R, n_generals = 2, n_groups = 6, nobs = 100,
         first_efa = first, second_efa = second)
}

}
\references{
Jiménez, M., Abad, F. J., Garcia-Garzon, E., & Garrido, L. E. (2023). Exploratory Bi-factor Analysis with Multiple General Factors. Multivariate behavioral research, 1–18. Advance online publication. https://doi.org/10.1080/00273171.2023.2189571
}
