% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bifactor.R
\name{summary.bifactor}
\alias{summary.bifactor}
\title{Compute summaries, fit indices, and information criteria from exploratory factor models.}
\usage{
## S3 method for class 'bifactor'
summary(bifactor, nobs=NULL, suppress=0, order=FALSE, ...)
summary.bifactor(bifactor, nobs=NULL, suppress=0, order=FALSE, ...)
}
\arguments{
\item{bifactor}{Object of class bifactor.}

\item{nobs}{Sample size. Defaults to NULL.}

\item{suppress}{Hide the loadings whose absolute magnitudes are smaller than this cutoff. Defaults to 0.}

\item{order}{Order the columns of the pattern matrix according to the variance they account for. Defaults to FALSE.}

\item{digits}{Number of digits to display in the loading and factor correlation matrices.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
List with the following objects:
\item{loadings}{Loading matrix.}
\item{Phi}{Factor correlation matrix.}
\item{communalities}{Communalities.}
\item{uniquenesses}{Uniquenesses.}
\item{complexity}{Item complexities.}
\item{VAR}{Variance accounted for the factors.}
\item{fit}{fit indices.}
}
\description{
Compute summaries, fit indices, and information criteria from exploratory factor models.
}
\details{
to be explained
}
\author{
Vithor R. Franco & Marcos Jiménez
}
