% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifactor.R
\name{top}
\alias{top}
\title{title: Square a number}
\usage{
top(mat, n = 5)
}
\arguments{
\item{x}{A numeric value to be squared}
}
\value{
The square of the input
}
\description{
Takes in any numeric value and squares it
}

