% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{random_oblq}
\alias{random_oblq}
\title{Generate random oblique matrices.}
\usage{
random_oblq(p, q)
}
\arguments{
\item{p}{Number of rows.}

\item{q}{Number of columns. Should not be greater than p.}
}
\value{
An oblique matrix with normally distributed data.
}
\description{
Generate random oblique matrices from a standard normal distribution.
}
\references{
Jiménez, M., Abad, F. J., Garcia-Garzon, E., & Garrido, L. E. (2023). Exploratory Bi-factor Analysis with Multiple General Factors. Multivariate behavioral research, 1–18. Advance online publication. https://doi.org/10.1080/00273171.2023.2189571
}
