% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{random_orth}
\alias{random_orth}
\title{Generate random orthogonal matrices.}
\usage{
random_orth(p, q)
}
\arguments{
\item{p}{Number of rows.}

\item{q}{Number of columns. Should not be greater than p.}
}
\value{
An orthogonal matrix with normally distributed data.
}
\description{
Generate random orthogonal matrices from a standard normal distribution. First, a matrix of random standard normal variables is simulated and then, the Q factor from the QR decomposition is returned.
}
\references{
Jiménez, M., Abad, F. J., Garcia-Garzon, E., & Garrido, L. E. (2023). Exploratory Bi-factor Analysis with Multiple General Factors. Multivariate behavioral research, 1–18. Advance online publication. https://doi.org/10.1080/00273171.2023.2189571
}
