% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{random_orth}
\alias{random_orth}
\title{Generate random orthogonal matrices.}
\usage{
random_orth(p, q)
}
\arguments{
\item{p}{Number of rows.}

\item{q}{Number of columns. Should not be greater than p.}
}
\value{
An orthogonal matrix with normally distributed data.
}
\description{
Generate random orthogonal matrices from a standard normal distribution. First, a matrix of random standard normal variables is simulated and then, the Q factor from the QR decomposition is returned.
}
\references{
Jiménez, M., Abad, F.J., Garcia-Garzon, E., Garrido, L.E. (2021, June 24). Exploratory Two-tier Modeling. Under review. Retrieved from https://osf.io/7aszj/?view_only=8f7bd98025104347a96f60a6736f5a64
}
