% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{random_poblq}
\alias{random_poblq}
\title{Generate a random partially oblique matrix.}
\usage{
random_poblq(p, q, oblq_factors)
}
\arguments{
\item{p}{Number of rows.}

\item{q}{Number of columns. Should not be greater than p.}

\item{oblq_factors}{A vector with the number of factors for each oblique block. E.g.: c(2, 4) means that there are two blocks of oblique factors: one with 2 factors and another with 4 factors. Everything else is orthogonal.}
}
\value{
A partially oblique matrix.
}
\description{
First, a matrix is simulated from a standard normal distribution. Second,the matrix is normalized and the Gram-Schmidt process is performed between the oblique blocks. Finally, the orthogonal blocks correspond to those columns of the Q matrix from the QR decomposition.
}
\examples{

random_poblq(p = 7, q = 7, oblq_factors = c(3, 2))

}
\references{
Jiménez, M., Abad, F. J., Garcia-Garzon, E., & Garrido, L. E. (2023). Exploratory Bi-factor Analysis with Multiple General Factors. Multivariate behavioral research, 1–18. Advance online publication. https://doi.org/10.1080/00273171.2023.2189571
}
