% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{retr_oblq}
\alias{retr_oblq}
\title{Retraction of a matrix onto the oblique manifold.}
\usage{
retr_oblq(X)
}
\arguments{
\item{X}{A matrix.}
}
\value{
An oblique matrix.
}
\description{
Transform a matrix into an oblique matrix.
}
\references{
Jiménez, M., Abad, F. J., Garcia-Garzon, E., & Garrido, L. E. (2023). Exploratory Bi-factor Analysis with Multiple General Factors. Multivariate behavioral research, 1–18. Advance online publication. https://doi.org/10.1080/00273171.2023.2189571
}
