% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{retr_orth}
\alias{retr_orth}
\title{Retraction of a matrix onto the orthogonal manifold.}
\usage{
retr_orth(X)
}
\arguments{
\item{X}{A matrix.}
}
\value{
An orthogonal matrix.
}
\description{
Transform a matrix into an orthogonal matrix.
}
\references{
Jiménez, M., Abad, F.J., Garcia-Garzon, E., Garrido, L.E. (2021, June 24). Exploratory Two-tier Modeling. Under review. Retrieved from https://osf.io/7aszj/?view_only=8f7bd98025104347a96f60a6736f5a64
}
