% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{retr_orth}
\alias{retr_orth}
\title{Retraction of a matrix onto the orthogonal manifold.}
\usage{
retr_orth(X)
}
\arguments{
\item{X}{A matrix.}
}
\value{
An orthogonal matrix.
}
\description{
Transform a matrix into an orthogonal matrix.
}
\references{
Jiménez, M., Abad, F. J., Garcia-Garzon, E., & Garrido, L. E. (2023). Exploratory Bi-factor Analysis with Multiple General Factors. Multivariate behavioral research, 1–18. Advance online publication. https://doi.org/10.1080/00273171.2023.2189571
}
