% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{se}
\alias{se}
\title{Standard errors for rotated factor loadings, factor correlations and uniquenesses.}
\usage{
se(fit = NULL, n = NULL, R = NULL, Lambda = NULL, Phi = NULL, X = NULL,
type = "normal", eta = 1)
}
\arguments{
\item{fit}{Optional \code{efast} model.}

\item{n}{Sample size.}

\item{X}{Raw data matrix.}

\item{type}{Type of random deviates: "normal", "elliptical" or "general".}

\item{eta}{Skewness parameter for elliptical data distributions.}
}
\value{
A list with the standard errors of the rotated factor loadings, factor correlations and uniquenesses.
}
\description{
Compute the sandwich standard errors of factor loadings, factor correlations and uniquenesses.
}
\details{
Currently, only available for \code{method = minres}.
}
\references{
Zhang G, Preacher KJ, Hattori M, Jiang G, Trichtinger LA (2019). A sandwich standard error estimator for exploratory factor analysis with nonnormal data and imperfect models. Applied Psychological Measurement, 43, 360–373. https://doi.org/10.1177/0146621618798669
}
