% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_factor.R
\name{sim_factor}
\alias{sim_factor}
\title{Simulate a bi-factor or generalized bifactor population structure.}
\usage{
sim_factor(n_generals, groups_per_general, items_per_group,
loadings_g = "medium", loadings_s = "medium",
crossloadings = 0, pure = FALSE,
generals_rho = 0, groups_rho = 0, confirmatory = TRUE,
method = "minres", fit = "rmsr", misfit = 0, error_method = "yuan")
}
\arguments{
\item{n_generals}{Number of general factors.}

\item{groups_per_general}{Number of group factors per general factor.}

\item{items_per_group}{Number of items per group factor.}

\item{loadings_g}{Loadings' magnitude on the general factors: "low", "medium" or "high". Defaults to "medium".}

\item{loadings_s}{Loadings' magnitude on the group factors: "low", "medium" or "high". Defaults to "medium".}

\item{crossloadings}{Magnitude of the cross-loadings among the group factors. Defaults to 0.}

\item{pure}{Fix a pure item on each general factor. Defaults to FALSE.}

\item{generals_rho}{Correlation among the general factors. Defaults to 0.}

\item{groups_rho}{Correlation among the group factors. Defaults to 0.}

\item{confirmatory}{Logical. Should the misfit value be computed according to a confirmatory model (TRUE) or an exploratory model (FALSE). Defaults to TRUE.}

\item{method}{Method used to generate population error: "minres" or "ml".}

\item{fit}{Fit index to control the population error.}

\item{misfit}{Misfit value to generate population error.}

\item{error_method}{Method used to control population error: c("yuan", "cudeck"). Defaults to "yuan".}
}
\value{
List with the following objects:
\item{lambda}{Population loading matrix.}
\item{Phi}{Population factor correlation matrix.}
\item{R}{Population correlation matrix.}
\item{R_error}{Population correlation matrix with error.}
\item{uniquenesses}{Vector of population uniquenesses.}
\item{delta}{Minimum of the objective function that correspond to the misfit value.}
}
\description{
Simulate a bi-factor or generalized bifactor population structure with cross-loading, pure items and correlated factors.
}
\details{
\code{sim_factor} generates bi-factor and generalized bifactor patterns with cross-loadings, pure items and
correlations among the general and group factors. When \code{crossloading} is different than 0, one cross-loading
is introduced for an item pertaining to each group factor. When \code{pure} is TRUE, one item loading of each group
factor is removed so that the item loads entirely on the general factor. To maintain the item communalities
constant upon these modifications, the item loading on the other factors may shrunk (if adding cross-loadings)
or increase (if setting pure items).

Loading magnitudes may range between 0.3-0.5 ("low"), 0.4-0.6 ("medium") and 0.5-0.7 ("high").
}
\references{
Jiménez, M., Abad, F.J., Garcia-Garzon, E., Garrido, L.E. (2021, June 24). Exploratory generalized bifactor Modeling. Under review. Retrieved from https://osf.io/7aszj/?view_only=8f7bd98025104347a96f60a6736f5a64
}
